+++ packages/node-core/src/db/db.module.test.ts
@@ -3,9 +3,13 @@
 
 import {INestApplication} from '@nestjs/common';
 import {Test} from '@nestjs/testing';
+import {ConfigureModule} from '@subql/node/src/configure/configure.module';
 import {Sequelize} from 'sequelize';
+import {NodeConfig} from '../configure/NodeConfig';
 import {SubqueryRepo} from '../entities';
 import {DbModule} from './db.module';
+
+const nodeConfig = new NodeConfig({subquery: 'packages/node-core/test/v1.0.0', subqueryName: 'test'});
 
 describe('DbModule', () => {
   let app: INestApplication;
@@ -17,13 +21,17 @@ describe('DbModule', () => {
   it('can connect to database', async () => {
     const module = await Test.createTestingModule({
       imports: [
-        DbModule.forRoot({
-          host: process.env.DB_HOST ?? '127.0.0.1',
-          port: process.env.DB_PORT ? Number(process.env.DB_PORT) : 5432,
-          username: process.env.DB_USER ?? 'postgres',
-          password: process.env.DB_PASS ?? 'postgres',
-          database: process.env.DB_DATABASE ?? 'postgres',
-        }),
+        ConfigureModule.testRegister(nodeConfig),
+        DbModule.forRootTest(
+          {
+            host: process.env.DB_HOST ?? '127.0.0.1',
+            port: process.env.DB_PORT ? Number(process.env.DB_PORT) : 5432,
+            username: process.env.DB_USER ?? 'postgres',
+            password: process.env.DB_PASS ?? 'postgres',
+            database: process.env.DB_DATABASE ?? 'postgres',
+          },
+          nodeConfig
+        ),
       ],
     }).compile();
 
@@ -36,13 +44,17 @@ describe('DbModule', () => {
   it('can load subquery model', async () => {
     const module = await Test.createTestingModule({
       imports: [
-        DbModule.forRoot({
-          host: process.env.DB_HOST ?? '127.0.0.1',
-          port: process.env.DB_PORT ? Number(process.env.DB_PORT) : 5432,
-          username: process.env.DB_USER ?? 'postgres',
-          password: process.env.DB_PASS ?? 'postgres',
-          database: process.env.DB_DATABASE ?? 'postgres',
-        }),
+        ConfigureModule.testRegister(nodeConfig),
+        DbModule.forRootTest(
+          {
+            host: process.env.DB_HOST ?? '127.0.0.1',
+            port: process.env.DB_PORT ? Number(process.env.DB_PORT) : 5432,
+            username: process.env.DB_USER ?? 'postgres',
+            password: process.env.DB_PASS ?? 'postgres',
+            database: process.env.DB_DATABASE ?? 'postgres',
+          },
+          nodeConfig
+        ),
         DbModule.forFeature(['Subquery']),
       ],
     }).compile();
