+++ packages/node-core/test/v1.0.0/distMock/chaintypes.js
@@ -1,0 +1,3 @@
+'use strict';
+
+exports.default = {typesBundle: {}};
