+++ packages/node-core/test/v1.0.0/distMock/index.js
@@ -1,0 +1,6 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', {value: true});
+const tslib_1 = require('tslib');
+//Exports all handler functions
+require('@polkadot/api-augment');
+(0, tslib_1.__exportStar)(require('./mappings/mappingHandlers'), exports);
