+++ packages/node-core/test/v1.0.0/schema.graphql
@@ -1,0 +1,11 @@
+type StarterEntity @entity {
+  id: ID! #id is a required field
+  field1: Int!
+
+  field2: String #filed2 is an optional field
+  field3: BigInt
+
+  field4: Date
+
+  field5: Boolean
+}
