+++ packages/node-core/test/v1.0.0/types.yaml
@@ -1,0 +1,28 @@
+types:
+  TestType: u32
+typesAlias:
+  Alias:
+    TestType2: test
+typesBundle:
+  spec:
+    '2312':
+      types:
+        - minmax:
+            - 232
+            - 122
+          types:
+            TestType3: test3
+  chain:
+    mockchain:
+      types:
+        - minmax:
+            - 232
+            - 122
+          types:
+            TestType4: test4
+typesChain:
+  chain2:
+    TestType5: test
+typesSpec:
+  spec3:
+    TestType6: test
